- share/mingle attribute is not copied by transfer_physical_appearance task - defer popopulate repo if export/import/transform fails - uint16_t errorCounter - abort import/export task messages

+++ include/utils/MiiUtils.h
@@ -19,14 +19,14 @@ namespace MiiUtils {
     time_t year_zero_offset(int year_zero); 
     uint32_t generate_timestamp(int year_zero, uint8_t ticks_per_sec); 
 
-    bool export_miis(uint8_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
-    bool import_miis(uint8_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
-    bool wipe_miis(uint8_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
+    bool export_miis(uint16_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
+    bool import_miis(uint16_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
+    bool wipe_miis(uint16_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
 
-    bool xform_miis(uint8_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
+    bool xform_miis(uint16_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
 
-    bool eight_fold_mii(uint8_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
-    bool copy_some_bytes_from_miis(uint8_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
+    bool eight_fold_mii(uint16_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
+    bool copy_some_bytes_from_miis(uint16_t &errorCounter, MiiProcessSharedState *mii_process_shared_state);
 
 
     void showMiiOperations(MiiProcessSharedState *mii_process_shared_state, size_t mii_index);
